pyproject.toml
[project]
name = "agent-evolve"
version = "0.1.0"
description = "Platform-agnostic evolutionary optimization system for codebases. AI agents cooperate to explore and evolve better solutions under human approval gates."
readme = "README.md"
authors = [
    { name = "Kyle", email = "kyleyhw@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "hypothesis>=6.115.0",
    "pygithub>=2.4.0",
    "jinja2>=3.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
agent-evolve = "agent_evolve.cli:main"

[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
